deeppresenter/deeppresenter/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deeppresenter"
authors = [{ name = "Hao Zheng", email = "wszh712811@gmail.com" }]
description = "An Agentic Framework for Reflective PowerPoint Generation"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dynamic = ["version"]


dependencies = [
    "aiohttp>=3.9.0",
    "arxiv>=2.2.0",
    "colorlog>=6.9.0",
    "docker>=7.1.0",
    "fake-useragent>=2.2.0",
    "fastmcp>=2.10.0,<2.14.0",
    "fasttext>=0.9.3",
    "filelock>=3.0.0",
    "firecrawl-py>=4.10.2",
    "gradio>=5.47.2,<6.0",
    "httpx>=0.28.0",
    "httpx-retries>=0.4.5",
    "jinja2>=3.1.6",
    "json-repair>=0.51.0",
    "jsonlines>=4.0.0",
    "langchain_mcp_adapters",
    "markitdown>=0.0.2",
    "mistune>=3.1.4",
    "numpy<2.0.0",
    "openai>=1.108.2",
    "platformdirs>=4.0.0",
    "playwright>=1.55.0",
    "pptagent>=0.2.18",
    "pre-commit>=4.3.0",
    "pydantic>=2.11.9",
    "pypdf>=6.1.1",
    "python-dotenv>=1.0.0",
    "semanticscholar>=0.11.0",
    "tavily-python>=0.7.14",
    "trafilatura>=2.0.0",
    "python-pptx>=0.6.21",
]

[project.urls]
Homepage = "https://github.com/Force1ess/DeepPresenter"
"Bug Tracker" = "https://github.com/Force1ess/DeepPresenter/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["deeppresenter*"]
where = ["."]

[tool.setuptools.package-data]
"deeppresenter" = [
    "roles/*.yaml",
    "html2pptx/*.js",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "llm: mark test as using LLM (requires OPENAI_API_KEY)",
]

[tool.ruff.lint]
ignore = ["F403", "F405", "E741", "E722"]

[tool.setuptools.dynamic]
version = { attr = "deeppresenter.__version__" }
